// packages/auth/src/types/auth.d.ts
export declare const UserSchema: z.ZodObject<{
    id: z.ZodString;
    email: z.ZodString;
    name: z.ZodString;
    avatarUrl: z.ZodOptional<z.ZodString>;
    emailVerified: z.ZodOptional<z.ZodBoolean>;
    isVerified: z.ZodOptional<z.ZodBoolean>;
    createdAt: z.ZodString;
    updatedAt: z.ZodString;
    role: z.ZodString;
}, "strip", z.ZodTypeAny, {
    name?: string;
    email?: string;
    id?: string;
    role?: string;
    avatarUrl?: string;
    emailVerified?: boolean;
    isVerified?: boolean;
    createdAt?: string;
    updatedAt?: string;
}, {
    name?: string;
    email?: string;
    id?: string;
    role?: string;
    avatarUrl?: string;
    emailVerified?: boolean;
    isVerified?: boolean;
    createdAt?: string;
    updatedAt?: string;
}>;
// ...
export declare const AuthStateSchema: z.ZodObject<{
    user: z.ZodNullable<z.ZodObject<{
        id: z.ZodString;
        email: z.ZodString;
        name: z.ZodString;
        avatarUrl: z.ZodOptional<z.ZodString>;
        emailVerified: z.ZodOptional<z.ZodBoolean>;
        isVerified: z.ZodOptional<z.ZodBoolean>;
        createdAt: z.ZodString;
        updatedAt: z.ZodString;
        role: z.ZodString;
    }, "strip", z.ZodTypeAny, {
        name?: string;
        email?: string;
        id?: string;
        role?: string;
        avatarUrl?: string;
        emailVerified?: boolean;
        isVerified?: boolean;
        createdAt?: string;
        updatedAt?: string;
    }, {
        name?: string;
        email?: string;
        id?: string;
        role?: string;
        avatarUrl?: string;
        emailVerified?: boolean;
        isVerified?: boolean;
        createdAt?: string;
        updatedAt?: string;
    }>>;
    tokens: z.ZodNullable<z.ZodObject<{
        accessToken: z.ZodString;
        refreshToken: z.ZodString;
        expiresIn: z.ZodNumber;
    }, "strip", z.ZodTypeAny, {
        accessToken?: string;
        refreshToken?: string;
        expiresIn?: number;
    }, {
        accessToken?: string;
        refreshToken?: string;
        expiresIn?: number;
    }>>;
    isAuthenticated: z.ZodBoolean;
    isLoading: z.ZodBoolean;
    isCheckingAuth: z.ZodOptional<z.ZodBoolean>;
    error: z.ZodNullable<z.ZodString>;
    message: z.ZodOptional<z.ZodNullable<z.ZodString>>;
}, "strip", z.ZodTypeAny, {
    user?: {
        name?: string;
        email?: string;
        id?: string;
        role?: string;
        avatarUrl?: string;
        emailVerified?: boolean;
        isVerified?: boolean;
        createdAt?: string;
        updatedAt?: string;
    };
    error?: string;
    message?: string;
    tokens?: {
        accessToken?: string;
        refreshToken?: string;
        expiresIn?: number;
    };
    isAuthenticated?: boolean;
    isLoading?: boolean;
    isCheckingAuth?: boolean;
}, {
    user?: {
        name?: string;
        email?: string;
        id?: string;
        role?: string;
        avatarUrl?: string;
        emailVerified?: boolean;
        isVerified?: boolean;
        createdAt?: string;
        updatedAt?: string;
    };
    error?: string;
    message?: string;
    tokens?: {
        accessToken?: string;
        refreshToken?: string;
        expiresIn?: number;
    };
    isAuthenticated?: boolean;
    isLoading?: boolean;
    isCheckingAuth?: boolean;
}>;
export interface User extends z.infer<typeof UserSchema> {
}
// ...
export interface AuthState extends z.infer<typeof AuthStateSchema> {
}